Create Unoptimized Setup
This commit sets up the unoptimized level. This level has multiple assets and lights within that drop performance. Objects and lights are set to be movable, ray tracing is enabled, and it will affect shadows.

The aim of this setup is to test the built-in light baking to test how it will affect performance overhead. At a baseline:

-render deferred lighting: 5.81ms
-shadow depths: 2.99ms
-lights: 5.04ms

-FPS: 46.5 (21.5ms)