Lab5 - TPA Foundations
* Create Unoptimized Setup

This commit sets up the unoptimized level. This level has multiple assets and lights within that drop performance. Objects and lights are set to be movable, ray tracing is enabled, and it will affect shadows.

The aim of this setup is to test the built-in light baking to test how it will affect performance overhead. At a baseline:

-render deferred lighting: 5.81ms
-shadow depths: 2.99ms
-lights: 5.04ms

-FPS: 46.5 (21.5ms)

* Re: Previous Commit

Had unsaved changes from previous commit. This commit adds the unsaved changes.

* Update Unoptimized Level

Have to resolve lost content. This commit updates the unoptimized level to include ray tracing and high shadow environments. Also, objects use meshes that have been subdivided a couple of times...

* Add Optimized World

* Add Optimization Strategies to Optimized Level

This commit adds optimized strategies to the level. Key targets were:

1) reducing casted shadows in the environment
2) Removing ray tracing
3) Reducing Level of detail. This was done by replacing the objects that had high triangle counts with their counterparts that had much lower triangle counts

* Create README.md

test video format

* Update README.md

Add video description

* Add Setup Description to README

* Add Results to README

* Restore Lab 5 README.md file

GitHub wanted to replace the README in Lab 5 with the README in the main branch