Add Additional Objects
This commit adds extra objects that I may play with in the scene. The goal of these objects is to limit test what happens when I add more objects of varying poly counts in a larger scene than previously. As usual, there are two variants: optimized and unoptimized, with the unoptimized version being subdivided multiple times to increase poly counts.