Lab6 performance at scale (#3)
* Add Additional Objects

This commit adds extra objects that I may play with in the scene. The goal of these objects is to limit test what happens when I add more objects of varying poly counts in a larger scene than previously. As usual, there are two variants: optimized and unoptimized, with the unoptimized version being subdivided multiple times to increase poly counts.

* Add Base Layout for Optimized World

Just breaking up commits so commit files stay light.

* Adding Pillars

Breaking up some commits to prevent large file sizes. This commit adds pillars to the optimized level

* Add Props

This commit adds props. The next commit will add the lighting so the lab can start.

* Add Lighting

This commit adds lighting to the scene. This marks the setup for the optimized level. The next commit will introduce the optimized level.

For base comparisons:
-renderedDeferredLighting: 0.99ms
-Lights: 0.13
-RTXDynamicGeometry: 0

* Remove Old Unoptimized Level

This commit removes the old unoptimized level. This commit is to break up commits so they do not grow too large. Next commit will duplicate the optimized world level and rename it to unoptimized world.

* Add Unoptimized World

This commit adds the unoptimized world. Future commits will continuously add unoptimized settings, and introduce higher-quality lighting and shadows.

* Replace optimized meshes with expensive mesh counterparts

* Add Lighting Detail, Shadows, & Ray Tracing

This commit finalizes the unoptimized level with better lighting quality. This marks the end of the unoptmized changes to the world. For Reference:
-RedneredDeferredLighting: 8.42
-Lights: 9.22
-RayTracingDynamicGeometry: 0.42